requirements.txt
transformers~=4.14.1
datasets~=1.17.0
OpenHowNet==1.0
click
conllu
editdistance
flaky
Flask
flask_cors
gensim
gevent
h5py
homoglyphs
language_tool_python
lemminflect
nbconvert
nbformat
nltk
numpy
numpydoc
overrides
pandas
Pillow
pyenchant
pyinflect
pytest
python_Levenshtein
requests
responses
scikit_learn
sentence_transformers
setuptools
spacy
Sphinx
sphinx_rtd_theme
strsimpy
tensorboardX
torch
Unidecode
anytree
text_visualization